Remove NDEF reading capabilities from entitlements.plist
Apple prohibits to put NDEF to entitlements.plist

On uploading it produces error:
ERROR ITMS-90778: "Invalid entitlement for core nfc framework.
The sdk version '13.0' and min OS version '13.0' are not compatible
for the entitlement '{com.apple.developer.nfc.readersession.formats}'
because '{NDEF is disallowed}'."

But this string is included by default. This commit deletes this key,
so app can be uploaded to TestFlight.